[MVP-38] refactor: 코드 리팩토링
