+++ src/main/java/somaMVP/service/FileService.java
@@ -1,36 +1,54 @@
 package somaMVP.service;
 
 import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
+import org.springframework.web.multipart.MultipartFile;
 import somaMVP.annotation.RunningTime;
+import somaMVP.response.ImageResponse;
+
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.*;
 import java.util.Base64;
 import java.util.List;
 
+import static somaMVP.service.ClearService.*;
+
 @Service
-@NoArgsConstructor
 @Slf4j
+@RequiredArgsConstructor
 public class FileService {
     @Value("${attachFileLocation}")
-    private String attachFileLocation;
-    @Value("${user.home}")
-    private String uploadDir;
+    public String attachFileLocation;
+    @Value("${HOME}")
+    public String uploadDir;
     private int imageNumber = 0;
+    private final ClearService clearService;
+    private final ImageResponse imageResponse;
     Path serverPath = Paths.get(
-            File.separator + "home" + File.separator + "ec2-user" +
+            File.separator + "home" + File.separator + "juwon" +
                     File.separator + // 다른 OS 환경 구분자 호환 용도
                     StringUtils.cleanPath("test.jpg"));
     @RunningTime
-    public void fileUpload(List<byte[]> multipartFile, List<String> originName) {
+    public void httpUpload(List<byte[]> multipartFile, List<String> originName) {
         for(byte[] imageBytes : multipartFile) {
             //String originNames = originName.get((imageNumber++)%loofCount); // Multipart 파일 이름 읽어올 때 사용
-            log.info("[{}번]{} -> {}", ++imageNumber, attachFileLocation, serverPath.toAbsolutePath());
             defaultUpload(imageBytes);
+            log.info("[{}번]{} -> {}", ++imageNumber, attachFileLocation, serverPath.toAbsolutePath());
+        }
+    }
+    @RunningTime
+    public void fileProcess(MultipartFile file) throws IOException {
+        multipartFiles.add(file.getInputStream().readAllBytes());
+        fileNames.add(file.getOriginalFilename());
+        if((imageResponse.sequenceNo) % FILE_ROOF_COUNT == 0){
+            httpUpload(multipartFiles, fileNames); // 파일 업로드 수행
+            clearService.clearList(multipartFiles, fileNames); // 업로드 후 리스트 초기화
         }
     }
     @RunningTime
@@ -55,6 +73,7 @@ public class FileService {
             Files.write(serverPath, imageBytes);
             log.info("path = {}", serverPath);
         } catch (IOException e) {
+            e.printStackTrace();
             log.info("IOException = {}", e.getMessage());
         }
     }
